elevaite_backend/toshiba_backend/requirements.txt
python-dotenv
pydantic
fastapi
uvicorn
qdrant-client
openai
requests
bs4
markdownify
matplotlib
cohere
scipy
cohere
pandas
scipy
sqlalchemy
asyncpg
aiohttp